build prerequisites: Remove libsdl2-dev again
The build issue with libsdl is nowadays generally solved by building it
in Yocto (libsdl-native) instead.  It was not necessary to include
libsdl2.

In fact, once the libsdl-native is used pervasively I imagine local
installation of libsdl1.2-dev is also not needed, but for the moment it
is still required by the standard Yocto documentation so we keep it.