Add SELinux support (Fedora, etc.)
The Makefile now executes the necessary chcon command to make the host
bind mounted directories for pipeline and yocto-cache storage accessible
from inside the container.

(FIXME: There is a :z and :Z flag introduced in docker that will do this
automatically instead.)

The command is disabled by default and needs to be uncommented by
SELinux users.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>