Add lavacli to standard install
It's a small tool and we might as well have it installed on all agents.

Later on it could also mean some optimizations are possible,
for example that test jobs are sent directly to lava-master from
any go-agent (i.e. the one that executed the build), rather than
having to redirect to specialized go-agent to set up lava jobs.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>