Update to v19.10.0
Both server and agent are updated.

For the server it is controlled by the base image specified in the
Dockerfile (and docker installation is the only supported way nowadays,
as far as this collection of scripts is concerned).

For the agent, a new version / download URL is provided in the
go-agent-setup.sh script (which is used by both the docker-based and
the native installation of agent)

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>